requirements.txt
alembic==1.7.7
click==8.1.3
colorama==0.4.4
Flask==2.1.2
Flask-Admin==1.6.0
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.0
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.11.4
install==1.3.5
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.0
MarkupSafe==2.1.1
PyJWT==2.4.0
python-dotenv==0.20.0
six==1.16.0
SQLAlchemy==1.4.36
Werkzeug==2.1.2
WTForms==3.0.1
WTForms-Appengine==0.1
zipp==3.8.0
